+++ src/modules/tsctl.c
@@ -64,8 +64,8 @@ CMD_FUNC(m_tsctl)
 
 	if (parv[1] && !stricmp(parv[1], "alltime"))
 	{
-		sendnotice(sptr, "*** Server=%s TStime=%li",
-			me.name, TStime());
+		sendnotice(sptr, "*** Server=%s TStime=%lld",
+			me.name, (long long)TStime());
 		sendto_server(cptr, 0, 0, NULL, ":%s TSCTL alltime", sptr->name);
 		return 0;
 	}
